+++ src/YLproxy.Infrastructure/Abstractions/IAppSettingsService.cs
@@ -1,0 +1,20 @@
+using YLproxy.Infrastructure;
+
+namespace YLproxy.Infrastructure.Abstractions;
+
+public interface IAppSettingsService
+{
+    T GetSection<T>(string sectionName) where T : class, new();
+    AppSettingsConfig GetConfig();
+
+
+
+
+
+
+
+
+
+    void Reload();
+}
+